June 20th @ 10:35 p.m.
Finished updating/making the file a lot cleaner, adding the 2020s, and removing extra variables. It was able to be knitted; it was .2 larger in storage, being 34.2 instead of 34. However, I think that is because of the new materai inlcuded at the end. Some more touch up is needed like turning off chunks and making sure the table of content is cleaner, but happy since I can start more on the demographics tomorrow.